style changes, add 404

+++ biography.py
@@ -3,7 +3,7 @@
 # TODO: Escape all reserved chars e.g. >, <, &, "
 
 import sys, os
-from pages import printBio
+from publisher.pages import printBio
 
 # bio print 
 # bio burn
@@ -16,18 +16,23 @@ PRINT = 'print'
 DIR_TO_BURN = '../public'
 
 def main():
+
+    if len(sys.argv) < 2:
+        return
+
     _, inst, *pages = sys.argv
     if inst == PRINT:
         printBio(pages)
         print("Finished printing pages!")
     elif inst == BURN:
         files = os.listdir(DIR_TO_BURN)
-        for file in filter(lambda x: x.endswith('.html'), files):
+        for file in filter(lambda x: x.endswith('.html') and x != '404.html', files):
             try: 
                 path = os.path.join(DIR_TO_BURN, file)
                 os.remove(path)
             except OSError:
                 print(f'Could not burn {file}')
+        
 
 
 main()
